requirements.txt
playwright>=1.49.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
allure-pytest>=2.13.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
